backend/requirements.txt
fastapi
uvicorn
langgraph
transformers
torch
psycopg2-binary
SQLAlchemy
pydantic
python-multipart
passlib[bcrypt]
python-jose
psycopg2